case when mail validate  token live time is ended

+++ backend/src/controllers/mail_varification/varification.ts
@@ -1,6 +1,8 @@
 import { ENV } from '@common/enums/app/app';
 import * as userSecurityService from '@services/mail_verification/user_security.service';
 
+import { validateMailToken } from '../../services/mail_verification/token.service';
+
 import type { TypedRequestQuery } from '@common/types/controller/controller';
 import type { Response, NextFunction } from 'express';
 
@@ -8,7 +10,7 @@ type EmailActivationRequestQuery = {
   link: string;
 };
 
-const activate = async (
+const activateMail = async (
   req: TypedRequestQuery<EmailActivationRequestQuery>,
   res: Response,
   next: NextFunction,
@@ -17,15 +19,19 @@ const activate = async (
     const userToken = req.params.link;
     const user = await userSecurityService.getByToken(userToken);
     if (!user) {
-      return res.redirect(ENV.MAIL.FAILED_URL);
+      return res.redirect(`${process.env.FRONTEND_URL}${ENV.MAIL.FAILED_URL}`);
+    }
+    const email = validateMailToken(userToken);
+    if (!email) {
+      return res.redirect(`${process.env.FRONTEND_URL}${ENV.MAIL.FAILED_URL}`);
     }
     const { id } = user;
     await userSecurityService.removeMailToken(id);
-    return res.redirect(ENV.MAIL.SUCCESS_URL);
+    return res.redirect(`${process.env.FRONTEND_URL}${ENV.MAIL.SUCCESS_URL}`);
   } catch (error) {
     console.error(error);
     next(error);
   }
 };
 
-export { activate };
+export { activateMail };
